Basic front, reservation does not work

+++ src/main/java/app/managers/CustomerManager.java
@@ -1,12 +1,12 @@
 package app.managers;
 
-import app.FunctionThrows;
 import app.dto.CustomerDTO;
 import app.exceptions.NotFoundException;
+import app.exceptions.UnauthorizedActionException;
 import app.model.Customer;
+import app.model.CustomerType;
 import app.repositories.CustomerRepository;
 import jakarta.inject.Inject;
-
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Predicate;
@@ -16,27 +16,36 @@ public class CustomerManager {
 	@Inject
 	private CustomerRepository customerRepository;
 
-	synchronized public CustomerDTO create (String username, String email, String password) throws Exception {
-		return new CustomerDTO(customerRepository.insert(new Customer(username, email, password)));
+	synchronized public CustomerDTO create (Customer customer) throws Exception {
+		return new CustomerDTO(customerRepository.insert(customer));
 	}
 
-	public void delete (int id) {
-		customerRepository.delete(id);
+	public Customer get (int id) throws NotFoundException, UnauthorizedActionException {
+		CustomerType type = this.customerRepository.get(id).getType();
+		if (type == CustomerType.CUSTOMER) {
+			return customerRepository.get(id);
+		} else {
+			throw new UnauthorizedActionException(type);
+		}
 	}
 
-	public Customer modify (int id, FunctionThrows<Customer> func) throws Exception {
-		return customerRepository.modify(id, func);
-	}
-
-	public Customer get (int id) throws NotFoundException {
-		return customerRepository.get(id);
+	synchronized public void delete (int id) throws Exception {
+		CustomerType type = this.customerRepository.get(id).getType();
+		if (this.customerRepository.get(id).getType() == CustomerType.CUSTOMER) {
+			customerRepository.delete(id);
+		} else {
+			throw new UnauthorizedActionException(type);
+		}
 	}
 
 	public Map<Integer, CustomerDTO> get (Predicate<Customer> predicate) {
 		try {
 			HashMap<Integer, CustomerDTO> res = new HashMap<>();
-			for (Map.Entry<Integer, Customer> entry : customerRepository.get(predicate).entrySet())
-				res.put(entry.getKey(), new CustomerDTO(entry.getValue()));
+			for (Map.Entry<Integer, Customer> entry : customerRepository.get(predicate).entrySet()) {
+				if (entry.getValue().getType() == CustomerType.CUSTOMER) {
+					res.put(entry.getKey(), new CustomerDTO(entry.getValue()));
+				}
+			}
 			return res;
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -45,14 +54,12 @@ public class CustomerManager {
 	}
 
 	public HashMap<Integer, CustomerDTO> getMap () {
-		HashMap<Integer, CustomerDTO> res = new HashMap<>();
-		for (Map.Entry<Integer, Customer> entry : customerRepository.getMap().entrySet())
-			res.put(entry.getKey(), new CustomerDTO(entry.getValue()));
-		return res;
+			HashMap<Integer, CustomerDTO> res = new HashMap<>();
+			for (Map.Entry<Integer, Customer> entry : customerRepository.getMap().entrySet()) {
+				if (entry.getValue().getType() == CustomerType.CUSTOMER) {
+					res.put(entry.getKey(), new CustomerDTO(entry.getValue()));
+				}
+			}
+			return res;
 	}
-
-	public int getLength () {
-		return customerRepository.getLenght();
-	}
-
 }
